+++ src/padrick/Generators/RTLGenerator/RTLGenerator.py
@@ -6,7 +6,8 @@ from typing import Tuple, Mapping
 
 import click_log
 import hjson
-from padrick.Generators.TemplateRenderJob import TemplateRenderJob
+from padrick.Generators.PadrickTemplate import PadrickTemplate
+from padrick.Generators.GeneratorSettings import RTLTemplates
 from padrick.Model.Padframe import Padframe
 from reggen import gen_rtl as reggen_gen_rtl
 from reggen import validate as reggen_validate
@@ -20,40 +21,21 @@ template_package = 'padrick.Generators.RTLGenerator.Templates'
 class RTLGenException(Exception):
     pass
 
-def generate_rtl(padframe: Padframe, dir: Path, header_text: str):
+def generate_rtl(templates: RTLTemplates, padframe: Padframe, dir: Path, header_text: str):
     os.makedirs(dir/"src", exist_ok=True)
     os.makedirs(dir/"include"/padframe.name, exist_ok=True)
-    TemplateRenderJob(name='SV package',
-                      target_file_name='pkg_{padframe.name}.sv',
-                      template=resources.read_text(template_package, 'pkg_padframe.sv.mako')
-                      ).render(dir/"src", logger=logger, padframe=padframe, header_text=header_text)
+    templates.toplevel_sv_package.render(dir/"src", logger=logger, padframe=padframe, header_text=header_text)
 
     next_pad_domain_reg_offset = 0 # Offset of the first register of the current pad_frame's register file. All
     address_ranges: Mapping[str, Tuple[int, int]] = {} # dictionary of pad_domain to start- end-address tupple
     # mappings. The end address is inclusive
     # registers in the padframe are mapped to a contiguous address space.
     for pad_domain in padframe.pad_domains:
-        TemplateRenderJob(name=f'Paddomain module {pad_domain.name}',
-                          target_file_name=f'{padframe.name}_{pad_domain.name}.sv',
-                          template=resources.read_text(template_package, 'pad_domain.sv.mako')
-                          ).render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain, header_text=header_text)
-        TemplateRenderJob(name=f'Pad instantiation module {pad_domain.name}',
-                          target_file_name=f'{padframe.name}_{pad_domain.name}_pads.sv',
-                          template=resources.read_text(template_package, 'pads.sv.mako')
-                          ).render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain, header_text=header_text)
-        TemplateRenderJob(name=f'Internal package for {pad_domain.name}',
-                          target_file_name=f'pkg_internal_{padframe.name}_{pad_domain.name}.sv',
-                          template=resources.read_text(template_package, 'pkg_pad_domain_internals.sv.mako')
-                          ).render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain, header_text=header_text)
-        TemplateRenderJob(name=f'Pad Multiplexer for {pad_domain.name}',
-                          target_file_name=f'{padframe.name}_{pad_domain.name}_muxer.sv',
-                          template=resources.read_text(template_package, 'pad_multiplexer.sv.mako')
-                          ).render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain,
-                                   debug_render=True, header_text=header_text)
-        TemplateRenderJob(name=f'Register File Specification for {pad_domain.name}',
-                          target_file_name=f'{padframe.name}_{pad_domain.name}_regs.hjson',
-                          template=resources.read_text(template_package, 'regfile.hjson.mako')
-                          ).render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain,
+        templates.pad_domain_top.render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain, header_text=header_text)
+        templates.pad_inst_module.render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain, header_text=header_text)
+        templates.internal_pkg.render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain, header_text=header_text)
+        templates.pad_mux_module.render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain, header_text=header_text)
+        templates.regfile_hjson.render(dir/"src", logger=logger, padframe=padframe, pad_domain=pad_domain,
                                    start_address_offset=hex(next_pad_domain_reg_offset), header_text=header_text)
 
 
@@ -73,28 +55,13 @@ def generate_rtl(padframe: Padframe, dir: Path, header_text: str):
             logger.error(f"Regtool template rendering of register file for pad domain {pad_domain.name} failed")
             raise RTLGenException("Reggen Rendering failed")
 
-    TemplateRenderJob(name='Padframe Top Module',
-                      target_file_name='{padframe.name}.sv',
-                      template=resources.read_text(template_package, 'padframe.sv.mako')
-                      ).render(dir / "src", logger=logger, padframe=padframe, address_ranges=address_ranges,
+    templates.toplevel_module.render(dir / "src", logger=logger, padframe=padframe, address_ranges=address_ranges,
                                address_space_size=address_space_size, header_text=header_text)
-    TemplateRenderJob(name='Padframe assignment header file',
-                      target_file_name='assign.svh',
-                      template=resources.read_text(template_package, 'assign.svh.mako')
-                      ).render(dir / "include" / padframe.name, logger=logger, padframe=padframe, header_text=header_text)
+    templates.assign_header_file.render(dir / "include" / padframe.name, logger=logger, padframe=padframe, header_text=header_text)
 
-    TemplateRenderJob(name=f'Bender.yml Project file',
-                      target_file_name="Bender.yml",
-                      template=resources.read_text(template_package, 'Bender.yml.mako')
-                      ).render(dir, logger=logger, padframe=padframe, header_text=header_text)
-    TemplateRenderJob(name=f'IPApprox src_files.yml',
-                      target_file_name="src_files.yml",
-                      template=resources.read_text(template_package, 'src_files.yml.mako')
-                      ).render(dir, logger=logger, padframe=padframe, header_text=header_text)
-    TemplateRenderJob(name=f'IPApprox ips_list.yml',
-                      target_file_name="ips_list.yml",
-                      template=resources.read_text(template_package, 'ips_list.yml.mako')
-                      ).render(dir, logger=logger, padframe=padframe, header_text=header_text)
+    templates.bender_project_file.render(dir, logger=logger, padframe=padframe, header_text=header_text)
+    templates.ipapprox_src_files_yml.render(dir, logger=logger, padframe=padframe, header_text=header_text)
+    templates.ipapprox_ips_list_yml.render(dir, logger=logger, padframe=padframe, header_text=header_text)